consolidate some silly components
