+++ app/notes/[slug]/page.tsx
@@ -1,14 +1,20 @@
 import * as runtime from "react/jsx-runtime";
+import { ErrorBoundary } from "react-error-boundary";
 import { evaluate } from "@mdx-js/mdx";
 import Content from "../../../components/Content";
-import PostMeta from "../../../components/PostMeta";
+import Link from "../../../components/Link";
+import Time from "../../../components/Time";
+import HitCounter from "../../../components/HitCounter";
 import Comments from "../../../components/Comments";
 import { getPostSlugs, getPostData } from "../../../lib/helpers/posts";
 import * as mdxComponents from "../../../lib/helpers/mdx-components";
 import { metadata as defaultMetadata } from "../../layout";
 import config from "../../../lib/config";
-import type { Metadata } from "next";
+import { FiCalendar, FiTag, FiEdit, FiEye } from "react-icons/fi";
+import type { Metadata, Route } from "next";
 import type { Article, WithContext } from "schema-dts";
+
+import styles from "./styles.module.css";
 
 // https://nextjs.org/docs/app/api-reference/functions/generate-static-params#disable-rendering-for-unspecified-paths
 export const dynamicParams = false;
@@ -94,7 +100,66 @@ export default async function Page({ params }: { params: Promise<{ slug: string 
     <>
       <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }} />
 
-      <PostMeta {...frontMatter} />
+      <div className={styles.meta}>
+        <div className={styles.item}>
+          <Link href={`/notes/${frontMatter.slug}` as Route} underline={false} className={styles.link}>
+            <FiCalendar className={styles.icon} />
+            <Time date={frontMatter.date} format="MMMM D, YYYY" />
+          </Link>
+        </div>
+
+        {frontMatter.tags && (
+          <div className={styles.item}>
+            <FiTag className={styles.icon} />
+            <span className={styles.tags}>
+              {frontMatter.tags.map((tag) => (
+                <span key={tag} title={tag} className={styles.tag} aria-label={`Tagged with ${tag}`}>
+                  {tag}
+                </span>
+              ))}
+            </span>
+          </div>
+        )}
+
+        <div className={styles.item}>
+          <Link
+            href={`https://github.com/${config.githubRepo}/blob/main/notes/${frontMatter.slug}.mdx`}
+            title={`Edit "${frontMatter.title}" on GitHub`}
+            underline={false}
+            className={styles.link}
+          >
+            <FiEdit className={styles.icon} />
+            <span>Improve This Post</span>
+          </Link>
+        </div>
+
+        {/* only count hits on production site */}
+        {process.env.NEXT_PUBLIC_VERCEL_ENV === "production" && (
+          <div
+            className={styles.item}
+            style={{
+              // fix potential layout shift when number of hits loads
+              minWidth: "7em",
+              marginRight: 0,
+            }}
+          >
+            {/* completely hide this block if anything goes wrong on the backend */}
+            <ErrorBoundary fallback={null}>
+              <FiEye className={styles.icon} />
+              <HitCounter slug={`notes/${frontMatter.slug}`} />
+            </ErrorBoundary>
+          </div>
+        )}
+      </div>
+
+      <h1 className={styles.title}>
+        <Link
+          href={`/notes/${frontMatter.slug}` as Route}
+          dangerouslySetInnerHTML={{ __html: frontMatter.htmlTitle || frontMatter.title }}
+          underline={false}
+          className={styles.link}
+        />
+      </h1>
 
       <Content>
         <MDXContent
