+++ app/projects/page.tsx
@@ -2,10 +2,12 @@ import { graphql } from "@octokit/graphql";
 import Content from "../../components/Content";
 import PageTitle from "../../components/PageTitle";
 import Link from "../../components/Link";
-import RepositoryCard from "../../components/RepositoryCard";
-import { SiGithub } from "react-icons/si";
+import RelativeTime from "../../components/RelativeTime";
+import commaNumber from "comma-number";
 import config from "../../lib/config";
 import { metadata as defaultMetadata } from "../layout";
+import { GoStar, GoRepoForked } from "react-icons/go";
+import { SiGithub } from "react-icons/si";
 import type { Metadata } from "next";
 import type { User, Repository } from "@octokit/graphql-schema";
 import type { Project } from "../../types";
@@ -101,7 +103,65 @@ export default async function Page() {
 
       <Content>
         <div className={styles.grid}>
-          {repos?.map((repo) => <RepositoryCard key={repo.name} className={styles.card} {...repo} />)}
+          {repos?.map((repo) => (
+            <div key={repo.name} className={styles.card}>
+              <Link
+                // @ts-ignore
+                href={repo.url}
+                className={styles.name}
+              >
+                {repo.name}
+              </Link>
+
+              {repo.description && <p className={styles.description}>{repo.description}</p>}
+
+              <div className={styles.meta}>
+                {repo.language && (
+                  <div className={styles.metaItem}>
+                    {repo.language.color && (
+                      <span className={styles.metaLanguage} style={{ backgroundColor: repo.language.color }} />
+                    )}
+                    {repo.language.name}
+                  </div>
+                )}
+
+                {repo.stars && repo.stars > 0 && (
+                  <div className={styles.metaItem}>
+                    <Link
+                      // @ts-ignore
+                      href={`${repo.url}/stargazers`}
+                      title={`${commaNumber(repo.stars)} ${repo.stars === 1 ? "star" : "stars"}`}
+                      underline={false}
+                      className={styles.metaLink}
+                    >
+                      <GoStar className={styles.metaIcon} />
+                      {commaNumber(repo.stars)}
+                    </Link>
+                  </div>
+                )}
+
+                {repo.forks && repo.forks > 0 && (
+                  <div className={styles.metaItem}>
+                    <Link
+                      // @ts-ignore
+                      href={`${repo.url}/network/members`}
+                      title={`${commaNumber(repo.forks)} ${repo.forks === 1 ? "fork" : "forks"}`}
+                      underline={false}
+                      className={styles.metaLink}
+                    >
+                      <GoRepoForked className={styles.metaIcon} />
+                      {commaNumber(repo.forks)}
+                    </Link>
+                  </div>
+                )}
+
+                {/* only use relative "time ago" on client side, since it'll be outdated via SSG and cause hydration errors */}
+                <div className={styles.metaItem}>
+                  <RelativeTime date={repo.updatedAt} verb="Updated" staticFormat="MMM D, YYYY" />
+                </div>
+              </div>
+            </div>
+          ))}
         </div>
 
         <p className={styles.viewMore}>
