--- components/Captcha/index.ts
@@ -1,2 +1,0 @@
-export * from "./Captcha";
-export { default } from "./Captcha";
