--- components/CodeHybrid/index.ts
@@ -1,2 +1,0 @@
-export * from "./CodeHybrid";
-export { default } from "./CodeHybrid";
