--- components/ContactForm/index.ts
@@ -1,2 +1,0 @@
-export * from "./ContactForm";
-export { default } from "./ContactForm";
