--- components/PostMeta/index.ts
@@ -1,2 +1,0 @@
-export * from "./PostMeta";
-export { default } from "./PostMeta";
