--- components/PostTitle/PostTitle.module.css
@@ -1,20 +1,0 @@
-.title {
-  margin: 0.3em 0 0.5em -1px; /* misaligned left margin, super nitpicky */
-  font-size: 2.1em;
-  line-height: 1.3;
-  font-weight: 700;
-}
-
-.title code {
-  margin: 0 0.075em;
-}
-
-.link {
-  color: var(--colors-text) !important;
-}
-
-@media (max-width: 768px) {
-  .title {
-    font-size: 1.8em;
-  }
-}
