--- components/PostTitle/index.ts
@@ -1,2 +1,0 @@
-export * from "./PostTitle";
-export { default } from "./PostTitle";
