--- components/PostsList/index.ts
@@ -1,2 +1,0 @@
-export * from "./PostsList";
-export { default } from "./PostsList";
