--- components/RepositoryCard/index.ts
@@ -1,2 +1,0 @@
-export * from "./RepositoryCard";
-export { default } from "./RepositoryCard";
