--- components/Selfie/Selfie.module.css
@@ -1,44 +1,0 @@
-.circleImage {
-  width: 50px;
-  height: 50px;
-  border: 1px solid var(--colors-light);
-  border-radius: 50%;
-}
-
-.link {
-  display: inline-flex;
-  flex-shrink: 0;
-  align-items: center;
-  color: var(--colors-mediumDark) !important;
-}
-
-.link:hover,
-.link:focus-visible {
-  color: var(--colors-link) !important;
-}
-
-.name {
-  margin: 0 0.6em;
-  font-size: 1.15em;
-  font-weight: 500;
-  letter-spacing: 0.02em;
-  line-height: 1;
-}
-
-
-@media (max-width: 768px) {
-  .circleImage {
-    width: 70px;
-    height: 70px;
-    border-width: 2px;
-  }
-
-  .link:hover .circleImage,
-  .link:focus-visible .circleImage {
-    borderColor: var(--colors-linkUnderline);
-  }
-
-  .name {
-    display: none;
-  }
-}
