--- components/Selfie/index.ts
@@ -1,2 +1,0 @@
-export * from "./Selfie";
-export { default } from "./Selfie";
