+++ lib/helpers/mdx-components.ts
@@ -6,7 +6,7 @@ export { default as Figure } from "../../components/Figure";
 
 // These (mostly very small) components are direct replacements for HTML tags generated by remark:
 export { default as a } from "../../components/Link";
-export { default as code } from "../../components/CodeHybrid";
+export { default as code } from "../../components/Code";
 export { default as blockquote } from "../../components/Blockquote";
 export { default as hr } from "../../components/HorizontalRule";
 export { H1 as h1, H2 as h2, H3 as h3, H4 as h4, H5 as h5, H6 as h6 } from "../../components/Heading";
