+++ package.json
@@ -21,7 +21,7 @@
   },
   "dependencies": {
     "@giscus/react": "^3.1.0",
-    "@libsql/client": "^0.8.1",
+    "@libsql/client": "0.15.0-pre.1",
     "@mdx-js/mdx": "^3.1.0",
     "@next/bundle-analyzer": "15.1.6",
     "@octokit/graphql": "^8.2.0",
@@ -65,7 +65,7 @@
     "remark-parse": "^11.0.0",
     "remark-rehype": "^11.1.1",
     "remark-smartypants": "^3.0.2",
-    "swr": "^2.3.1",
+    "swr": "^2.3.2",
     "unified": "^11.0.5"
   },
   "devDependencies": {
